Yield sort of works with revel, but mostly not
When revel is setting up your app to run, it looks for structs that
embed github.com/robfig/revel.Controller and uses that to find
controllers. As it doesn't check outside the current package for types
embedding revel.Controller, it doesn't find controllers that use the
embedded yield.Controller. I have a version that checks for resemblence
to the Controller, instead of of exact matching and that works, but I
need to look into the best way to deal with that.